pyproject.toml
[project]
name = "uv-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "chromadb-client>=1.3.6",
    "cryptography>=46.0.3",
    "duckduckgo-search<7",
    "fastapi>=0.127.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.35",
    "langgraph>=0.5.4",
    "langgraph-checkpoint-redis>=0.1.3",
    "mem0ai>=1.0.1",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "psycopg2-binary>=2.9.9",
    "modelscope>=1.10.0",
    "datasets>=2.16.0,<3.0.0",
    "addict>=2.4.0",
    "sqlmodel>=0.0.14",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.41",
    "tabulate>=0.9.0",
    "uvicorn>=0.40.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sentence-transformers>=5.2.0",
    "arize-phoenix>=12.27.0",
    "openinference-instrumentation-langchain>=0.1.57",
]

[tool.uv]
#python-install-mirror = "https://mirror.nju.edu.cn/github-release/indygreg/python-build-standalone"  南京大学（已废弃）
python-install-mirror = "https://registry.npmmirror.com/-/binary/python-build-standalone/"   # 阿里源

